Add simple rate-limiting to emacswiki downloads
For the sync case, we set the next allowed download time to the
current time plus two seconds after each emacswiki download. Then,
before starting an emacs wiki download, we check if it is after that
time, and if not, we wait until after it.

For the async case, I've implemented a queue that waits 2 seconds in
between running items.

The sync case seems to be working ok. You can test it by doing:

    test/test-recipe.sh notify

The async case is untested as of yet