Remove wait mode from el-get
It didn't actually work right. I waits for all packages to be in
status "installed" or "required", but the status of each package is
set to "required" at the beginning of the install, so the wait loop
just ends immediately, and it effectively just acts like async mode.

TODO: Update documentation.